+++ GCP_Raytracer_Framework/Camera.h
@@ -5,7 +5,7 @@ class Camera
 {
 public:
 	Camera(glm::ivec2 winsize);
-	Ray GetRay(glm::ivec2 windowPos,glm::ivec2 winSize);
+	Ray GetRay(glm::vec2 windowPos,glm::ivec2 winSize);
 private:
 	glm::mat4 mViewMat;
 	glm::mat4 mProjMat;
